Babel Setup - Getting Stuck and Solving with Babel Errors
Installing Babel and Configuring it.
This is a 3 step process.
Step 1 - Installing the below mentioned packages. (4 of them) and adding them in the webpack.config.js in 'entry' and 'modules' with babel loader using regular statements ( / *regular statement* / )
Step 2 - Create a babel config file ( .babelrc ) to tell babel which stuff do we want to convert to ES5 by targetting browser versions
Step 3 - Installing polyfills package to polyfill the newly intoduced features in
ES6 which can't be converted. We also have to add it in the entry points in the webpack.config.js file
Installed Packages Commands:
- babel-loader
- babel-preset-env
- babel-core
- babel-polyfill

Other than the 'babel-loader' all the other dependencies were obselete versions and would not work with the current version of babel-loader...
So to tackle the problem, I had to install the new versions of the packages with all new npm commands as shown below.
- @babel/core
- @babel/preset-env
- @babel/polyfill
These packages are installed as new packages in the package.json instead of just updating the existing versions. This maybe due to the different syntax in their names.
To differentiate between these packages i've added a line space between these new packages and the rest of packages in the package.json file.

Use of packages:
Polyfill:
  There were some new features in ES6 that didn't even exist in any form in the previous versions like promises object. So these features must be polyfilled (added) into the final js file and for that we make use of this package. Also, as the code is going to be included in our code files, so we will install it as a dependency and not as a dev dependency.